third_party/blink/web_tests/platform/mac-mac12/external/wpt/webrtc/RTCRtpSender-getStats.https-expected.txt

This is a testharness.js-based test.
FAIL sender.getStats() via addTransceiver should return stats report containing outbound-rtp stats assert_equals: Expect dictionary.senderId to be string expected "string" but got "undefined"
FAIL sender.getStats() via addTrack should return stats report containing outbound-rtp stats assert_equals: Expect dictionary.senderId to be string expected "string" but got "undefined"
FAIL sender.getStats() should return stats report containing ICE candidate stats assert_true: Expect statsReport to contain stats object of type candidate-pair expected true got false
Harness: the test ran to completion.
